// src/app/interfaces/client.ts
export interface Client {
    id?: string
    Nom_Client?: string
    Pays?: string
    Ville?: string
    Region?: string
    Telephone?: string
    Email?: string
    Categorie_Tarifaire?: string
    Adresse?: string
    Id_Commercial?: number
    Nom_Commercial?: string
    Code_Commercial?: number
}
 